scattered  the huge zabbix.py file
